mb/google/poppy/variants/nocturne: sx9310 to 400kb
The spec of the sx9310 says the I2C interface can handle standard
(100kb/s) and fast mode (400kb/s).  The current setting is using fast
plus (1000kb/s) so this change is reducing the speed to fast mode.

I've been using the sensors with this change for a few weeks now, though
I also don't recall seeing an issue prior to this change.

Change-Id: I337fc02c52565d6ec4d7bac1b3564f65238962dc
Signed-off-by: Matt Delco <delco@chromium.org>
Reviewed-on: https://review.coreboot.org/28075
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Furquan Shaikh <furquan@google.com>