amd/stoneyridge: Remove SATA D0 on suspend
Remove the step of setting the SATA controller to S0 as the system is
entering S3.  This had been duplicated from AMD's FchCarrizo.asl file,
but upon closer inspection, the conditions for this step to run cannot
be met.  This does not affect Grunt's behavior, as the SATA controller
is disabled.

TEST=Suspend and resume Grunt
BUG=b:77602074

Change-Id: Ib269a5363d03c7048abd0c8a9a28df92a773790c
Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Reviewed-on: https://review.coreboot.org/28863
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Martin Roth <martinroth@google.com>