MAINTAINERS: Clarify this is about active upstream development
<vendor> seems to be confused about the meaning of our maintainers list.
I get the feeling some use it to organize corporate internal teams and
branches, adding names to the list that don't show up in Gerrit and even
if, often don't react to reviewing requests (within months). Maybe they
even don't know that this is about coreboot.org?

To clarify this:

  o Add an introductory paragraph mentioning development on coreboot.org.
  o Explicitly state that maintainers should be registered to Gerrit.
  o If a topic is tagged as `Supported` or `Maintained`, expect that
    somebody reacts to review requests.

Change-Id: I9ee038dc5ee1f4993ba1d230ef6e737f20e2ff8a
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.coreboot.org/29471
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-by: David Guckian